Add viem to import map, combine abi utils
