+++ web/routes/api/sources.ts
@@ -4,11 +4,11 @@ import { type Handlers, Status } from "fresh/server.ts";
 import type { WithSession } from "fresh-session";
 
 import { Buffer } from "node:buffer";
-import { getAddress, toBytes, toHex } from "npm:viem";
+import { getAddress, toBytes, toHex } from "viem";
 
 import { amqpChannel, prisma } from "web/main.ts";
 import { checkPermission, logRequest } from "web/util.ts";
-import { formatAbiItemPrototype } from "~/abitype.ts";
+import { formatAbiItemPrototype } from "~/utils/abiUtils.ts";
 import { reload as reloadControl } from "~/messages/control.ts";
 import { ControlObserverRoutingKey } from "~/constants/constants.ts";
 import type { User } from "~/prisma/shim.ts";
