+++ web/routes/api/sources/testWebhook.ts
@@ -3,7 +3,7 @@ import { LogLevels } from "std/log/levels.ts";
 import { type Handlers, Status } from "fresh/server.ts";
 
 import { Buffer } from "node:buffer";
-import { toBytes } from "npm:viem";
+import { toBytes } from "viem";
 
 import { amqpChannel, prisma } from "web/main.ts";
 import { logRequest } from "web/util.ts";
